Some new slashes

+++ scommands/mod/purge.js
@@ -1,9 +1,10 @@
-const { Collection, MessageEmbed, MessageActionRow, MessageButton, MessageAttachment } = require('discord.js');
+const { CommandInteraction } = require('discord.js');
 
 module.exports = {
     name: 'purge',
     description: 'delete x amount of messages',
-    permission: ['MANAGE_MESSAGES'],
+    userperm: ['MANAGE_MESSAGES'],
+    botperm: ['MANAGE_MESSAGES'],
     ownerOnly: false,
     options: [
         {
@@ -14,17 +15,17 @@ module.exports = {
         }
     ],
     /** 
-     * @param {Message} message 
+     * @param {CommandInteraction} interaction 
      * @param {String[]} args 
      */
-    run: async (bot, interaction, args, message) => {
-        let amount = args[0]
-        if (amount <= 100) {
-            interaction.channel.bulkDelete(amount, true)
-        }
-
-        interaction.channel.send({
-            content: `I've cleared \`${amount}\` messages :broom:`
-        })
-    }
-}
+    run: async (bot, interaction, args) => {
+     
+let amount = args[0]        
+ if (amount < 2) return interaction.followUp({content: 'You cannot delete less than 2 messages'});
+if (amount <= 100) {            
+    interaction.channel.bulkDelete(amount, true)       
+         }
+      
+        interaction.channel.send({            content: `I've cleared \`${amount}\` messages :broom:`       
+})   
+    }}
